Обновлён TemplateBuilder
